timber.go: Avoid closing the record channel on shutdown
The current method of closing down the record channel is racy. The
blackhole flag is set but writers can miss that and send to the record
channel after it has been closed resulting in a panic. There's no easy
fix for this, but the solution here is to not close the record channel
and do a best effort attempt to drain it.

This fix has a couple of downsides:

1. Log messages may get left in the record channel. This isn't a big
problem as they were written in a racy manner and could easily have hit
the blackhole instead.

2. If the record channel fills up (current capacity is 300) after close
but the best effort draining does not clear it out then subsequent logs
may block. It's possible we could fix this by doing a simultaneous read
on the blackhole channel but in any case the window for this happening
should be very short.